@@ -12,52 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "8.2.12",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
+    "@angular/animations": "13.1.3",
+    "@angular/cdk": "^13.1.3",
+    "@angular/common": "13.1.3",
+    "@angular/compiler": "13.1.3",
+    "@angular/core": "13.1.3",
+    "@angular/flex-layout": "^13.0.0-beta.36",
+    "@angular/forms": "13.1.3",
+    "@angular/material": "^13.1.3",
+    "@angular/platform-browser": "13.1.3",
+    "@angular/platform-browser-dynamic": "13.1.3",
+    "@angular/platform-server": "13.1.3",
+    "@angular/router": "13.1.3",
     "bootstrap": "^4.6.0",
     "core-js": "^3.8.3",
     "jquery": "^3.5.1",
-    "node-sass": "^4.14.1",
+    "node-sass": "^7.0.1",
     "oidc-client": "^1.11.3",
     "popper.js": "^1.16.0",
-    "protractor": "~5.4.2",
     "rxjs": "^6.6.3",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "zone.js": "0.9.1"
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "^13.1.4",
+    "@angular/cli": "^13.1.4",
+    "@angular/compiler-cli": "^13.1.3",
+    "@angular/language-service": "^13.1.3",
     "@types/jasmine": "~3.4.4",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.2",
+    "codelyzer": "^6.0.2",
     "ini": "^1.3.7",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
+    "jasmine-core": "~4.0.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "^6.3.11",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.5.3"
+    "typescript": "4.5.4"
   },
   "optionalDependencies": {
-    "protractor": "~5.4.2",
+    "protractor": "~7.0.0",
     "ts-node": "~8.4.1",
     "tslint": "~5.20.0"
   }
