@@ -1,11 +1,11 @@
-import { NgModule } from '@angular/core';
-import { ServerModule } from '@angular/platform-server';
-import { ModuleMapLoaderModule } from '@nguniversal/module-map-ngfactory-loader';
-import { AppComponent } from './app.component';
-import { AppModule } from './app.module';
+import {NgModule} from '@angular/core';
+import {ServerModule} from '@angular/platform-server';
+import {AppComponent} from './app.component';
+import {AppModule} from './app.module';
 
 @NgModule({
-    imports: [AppModule, ServerModule, ModuleMapLoaderModule],
-    bootstrap: [AppComponent]
+  imports: [AppModule, ServerModule],
+  bootstrap: [AppComponent]
 })
-export class AppServerModule { }
+export class AppServerModule {
+}
